+++ src/components/Search/Search.module.scss
@@ -1,17 +1,9 @@
 @import '../../styles/vars.scss';
 
-.container {
+.form {
   display: flex;
-  flex: 1 1;
-  flex-direction: column;
-  align-items: center;
-  padding: 0px 100px 90px 100px;
-
-  .form {
-    display: flex;
-    height: 56px;
-    margin: 30px 0px;
-  }
+  height: 56px;
+  margin: 30px 0px;
 }
 
 .locationInput {
@@ -121,28 +113,4 @@
   color: $flamingo;
   font-size: 14px;
   line-height: 19px;
-}
-
-.pageContent {
-  flex: 1 1;
-  padding: 100px;
-  border-radius: 30px;
-  background: url('../../assets/background.jpg') center;
-  background-size: 100% 100%;
-  background-repeat: no-repeat;
-  color: white;
-
-  h1 {
-    margin-bottom: 17px;
-    font-weight: bold;
-    font-size: 48px;
-    line-height: 56px;
-  }
-
-  span {
-    display: block;
-    max-width: 50%;
-    font-size: 18px;
-    line-height: 27px;
-  }
 }
