+++ src/containers/Home/Home.module.scss
@@ -1,0 +1,31 @@
+.container {
+  display: flex;
+  flex: 1 1;
+  flex-direction: column;
+  align-items: center;
+  padding: 0px 100px 90px 100px;
+}
+
+.pageContent {
+  flex: 1 1;
+  padding: 100px;
+  border-radius: 30px;
+  background: url('../../assets/background.jpg') center;
+  background-size: 100% 100%;
+  background-repeat: no-repeat;
+  color: white;
+
+  h1 {
+    margin-bottom: 17px;
+    font-weight: bold;
+    font-size: 48px;
+    line-height: 56px;
+  }
+
+  span {
+    display: block;
+    max-width: 50%;
+    font-size: 18px;
+    line-height: 27px;
+  }
+}
