+++ src/containers/index.js
@@ -1,1 +1,1 @@
-export * from './Home';
+export * from './Home/Home';
