--- src/routes/index.js
@@ -1,3 +1,0 @@
-import {
-  Home
-} from '../containers'
